% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_episcore.R
\name{calculate_episcore}
\alias{calculate_episcore}
\title{Calculate a single episcore}
\usage{
calculate_episcore(
  thrs.criteria = 0.05,
  beta.file = "methylscore/betamatrix_test.txt",
  ewas.file = "methylscore/sumstats_test.txt",
  missingness = 0.2,
  path = getwd()
)
}
\arguments{
\item{thrs.criteria}{Thresholding criteria for CpG selection.}

\item{ewas.file}{Path and name to the EWAS file.}

\item{missingness}{Maximum tolerated missing data for CpG sites.}

\item{beta}{Beta from sample CpG.}
}
\value{
The calculated episcore.
}
\description{
This function calculates an epigenetic score based on methylation data. Two inputs are required: a beta matrix with CpG methylation and an external reference with the association and effect values for each CpG
}
\details{
This function calculates one epigenetic score using methylation information.
}
