% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multiple_episcores.R
\name{calculate_multiple_episcores}
\alias{calculate_multiple_episcores}
\title{Calculate multiple episcores}
\usage{
calculate_multiple_episcores(
  thrs.criteria = 0.05,
  beta.file = "",
  ewas.list.path = "",
  missingness = 0.2
)
}
\arguments{
\item{thrs.criteria}{Thresholding criteria for CpG selection.}

\item{missingness}{Maximum tolerated missing data for CpG sites.}

\item{beta}{A preloaded beta matrix (data.frame) with CpG methylation values.}

\item{ewas.file}{Path to the directory of EWAS files.}

\item{path}{Directory to save log files.}
}
\value{
A data.frame with calculated episcores for all EWAS files.
}
\description{
This function calculates multiple epigenetic score based on methylation data. It requires a beta matrix and a directory containing EWAS files.
}
\details{
This function calculates multiple epigenetic scores using methylation information.
}
