% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multiple_episcores.R
\name{calculate_multiple_episcores}
\alias{calculate_multiple_episcores}
\title{Calculate multiple episcores}
\usage{
calculate_multiple_episcores(
  thrs.criteria = 0.05,
  beta.file = "methylscore/betamatrix_test.txt",
  ewas.list.path = "methylscore/sumstats_test_multiple/",
  missingness = 0.2,
  path = getwd()
)
}
\arguments{
\item{thrs.criteria}{Thresholding criteria for CpG selection.}

\item{missingness}{Maximum tolerated missing data for CpG sites.}

\item{beta}{Beta from sample CpG.}

\item{ewas.file}{Path and name to the list of EWAS files.}
}
\value{
The calculated episcore.
}
\description{
This function calculates multiple epigenetic score based on methylation data. Two inputs are required: a beta matrix with CpG methylation and a list of external references with the association and effect values for each CpG
}
\details{
This function calculates multiple epigenetic scores using methylation information.
}
